move diagnostics plots to eccDefinition, make freqFir derive from eccDefinition
